+++ src/VirtualDesktopIndicator/Api/IVirtualDesktop.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace VirtualDesktopIndicator.Api
+{
+    interface IVirtualDesktopApi
+    {
+        uint Current();
+    }
+}
