+++ src/VirtualDesktopIndicator/VirtualDesktopIndicator.csproj
@@ -110,13 +110,14 @@
     <Reference Include="System.Windows.Forms" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Api\IVirtualDesktop.cs" />
+    <Compile Include="Api\Latest.cs" />
     <Compile Include="Helpers\AutorunManager.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Helpers\SystemMetricsApi.cs" />
     <Compile Include="Helpers\RegistryMonitor.cs" />
     <Compile Include="TrayIndicator.cs" />
-    <Compile Include="Helpers\VirtualDesktopApi.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
